refactor: move all draggable menus/configurators to respective files
