+++ .vscode/settings.json
@@ -64,4 +64,10 @@
     "licenser.license": "GPLv3",
     "licenser.projectName": "Autosteering",
     "licenser.disableAutoHeaderInsertion": false,
+    "explorer.fileNesting.patterns": {
+        "pubspec.yaml": "pubspec.lock,pubspec_overrides.yaml,.packages,.flutter-plugins,.flutter-plugins-dependencies,.metadata",
+        "*.dart": "${capture}.g.dart, ${capture}.freezed.dart",
+    },
+    "explorer.fileNesting.enabled": true,
+    "explorer.fileNesting.expand": false,
 }
