+++ lib/src/features/guidance/providers/autosteering_parameter_providers.dart
@@ -15,6 +15,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Autosteering.  If not, see <https://www.gnu.org/licenses/>.
 
+import 'dart:ui';
+
+import 'package:autosteering/src/features/common/common.dart';
+import 'package:autosteering/src/features/settings/settings.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'autosteering_parameter_providers.g.dart';
@@ -31,4 +35,33 @@ class ShowAutosteeringParameterConfig
 
   /// Invert the current [state].
   void toggle() => Future(() => state = !state);
+}
+
+/// A provider for the UI [Offset] for the autosteering configurator.
+@riverpod
+class AutosteeringConfiguratorUiOffset
+    extends _$AutosteeringConfiguratorUiOffset {
+  @override
+  Offset build() {
+    ref
+      ..watch(reloadAllSettingsProvider)
+      ..listenSelf((previous, next) {
+        if (previous != null && next != previous) {
+          ref.read(settingsProvider.notifier).update(
+                SettingsKey.uiAutosteeringConfiguratorOffset,
+                next.toJson(),
+              );
+        }
+      });
+
+    final setting = ref
+        .read(settingsProvider.notifier)
+        .getMap(SettingsKey.uiAutosteeringConfiguratorOffset);
+    return setting != null
+        ? OffsetJsonExtension.fromJson(Map<String, dynamic>.from(setting))
+        : Offset.zero;
+  }
+
+  /// Updates [state] to [value].
+  void update(Offset value) => Future(() => state = value);
 }
