+++ lib/src/features/guidance/widgets/autosteering_parameter_configurator.dart
@@ -16,6 +16,8 @@
 // along with Autosteering.  If not, see <https://www.gnu.org/licenses/>.
 
 import 'dart:async';
+import 'dart:math';
+import 'dart:ui';
 
 import 'package:autosteering/src/features/common/common.dart';
 import 'package:autosteering/src/features/guidance/guidance.dart';
@@ -95,7 +97,6 @@ class AutosteeringParameterConfigurator extends StatelessWidget {
     );
   }
 }
-
 
 class _PurePursuitConfigurator extends ConsumerWidget {
   const _PurePursuitConfigurator();
@@ -364,4 +365,95 @@ class _StanleyParametersConfigurator extends ConsumerWidget {
       ],
     );
   }
+}
+
+/// A draggable version of [AutosteeringParameterConfigurator], typically used
+/// as a child of a [Stack] that is a child of a [LayoutBuilder].
+
+class DraggableAutosteeringParameterConfigurator
+    extends ConsumerStatefulWidget {
+  /// A draggable version of [AutosteeringParameterConfigurator], typically used
+  /// as a child of a [Stack] that is a child of a [LayoutBuilder].
+  ///
+  /// [constraints] are used to layout the widget.
+  const DraggableAutosteeringParameterConfigurator({
+    required this.constraints,
+    super.key,
+  });
+
+  /// Constraints used to layout this widget.
+  final BoxConstraints constraints;
+
+  @override
+  ConsumerState<ConsumerStatefulWidget> createState() =>
+      _DraggableAutosteeringParameterConfiguratorState();
+}
+
+class _DraggableAutosteeringParameterConfiguratorState
+    extends ConsumerState<DraggableAutosteeringParameterConfigurator> {
+  late Offset offset = ref.read(autosteeringConfiguratorUiOffsetProvider);
+
+  @override
+  Widget build(BuildContext context) => Positioned(
+        left: clampDouble(
+          offset.dx,
+          0,
+          widget.constraints.maxWidth - 380,
+        ),
+        top: clampDouble(offset.dy, 0, widget.constraints.maxHeight - 390),
+        child: Padding(
+          padding: const EdgeInsets.all(8),
+          child: SizedBox(
+            height: min(
+              385,
+              widget.constraints.maxHeight -
+                  clampDouble(
+                    offset.dy,
+                    0,
+                    widget.constraints.maxHeight - 390,
+                  ),
+            ),
+            child: LongPressDraggable(
+              onDragUpdate: (update) {
+                setState(
+                  () => offset = Offset(
+                    offset.dx + update.delta.dx,
+                    offset.dy + update.delta.dy,
+                  ),
+                );
+              },
+              onDragEnd: (details) => ref
+                  .read(
+                    autosteeringConfiguratorUiOffsetProvider.notifier,
+                  )
+                  .update(
+                    Offset(
+                      clampDouble(
+                        offset.dx,
+                        0,
+                        widget.constraints.maxWidth - 380,
+                      ),
+                      clampDouble(
+                        offset.dy,
+                        0,
+                        widget.constraints.maxHeight - 390,
+                      ),
+                    ),
+                  ),
+              childWhenDragging: const SizedBox.shrink(),
+              feedback: const Opacity(
+                opacity: 0.7,
+                child: SizedBox(
+                  height: 385,
+                  child: AutosteeringParameterConfigurator(),
+                ),
+              ),
+              child: const SizedBox(
+                height: 385,
+                child: AutosteeringParameterConfigurator(),
+              ),
+            ),
+          ),
+        ),
+      );
 }
