+++ lib/src/features/guidance/widgets/path_recording_menu.dart
@@ -16,6 +16,8 @@
 // along with Autosteering.  If not, see <https://www.gnu.org/licenses/>.
 
 import 'dart:async';
+import 'dart:math';
+import 'dart:ui';
 
 import 'package:autosteering/src/features/field/field.dart';
 import 'package:autosteering/src/features/guidance/guidance.dart';
@@ -661,4 +663,84 @@ class _CreateFieldButton extends ConsumerWidget {
       ),
     );
   }
+}
+
+/// A draggable version of [PathRecordingMenu], typically used as
+/// a child of a [Stack] that is a child of a [LayoutBuilder].
+class DraggablePathRecordingMenu extends ConsumerStatefulWidget {
+  /// A draggable version of [PathRecordingMenu], typically used as
+  /// a child of a [Stack] that is a child of a [LayoutBuilder].
+  ///
+  /// [constraints] are used to layout the widget.
+  const DraggablePathRecordingMenu({required this.constraints, super.key});
+
+  /// Constraints used to layout this widget.
+  final BoxConstraints constraints;
+  @override
+  ConsumerState<ConsumerStatefulWidget> createState() =>
+      _DraggablePathRecordingMenuState();
+}
+
+class _DraggablePathRecordingMenuState
+    extends ConsumerState<DraggablePathRecordingMenu> {
+  late Offset offset = ref.read(pathRecordingMenuUiOffsetProvider);
+
+  @override
+  Widget build(BuildContext context) => Positioned(
+        left: clampDouble(
+          offset.dx,
+          0,
+          widget.constraints.maxWidth - 380,
+        ),
+        top: clampDouble(offset.dy, 0, widget.constraints.maxHeight - 350),
+        child: Padding(
+          padding: const EdgeInsets.all(8),
+          child: SizedBox(
+            height: min(
+              700,
+              widget.constraints.maxHeight -
+                  clampDouble(
+                    offset.dy,
+                    0,
+                    widget.constraints.maxHeight - 350,
+                  ),
+            ),
+            child: LongPressDraggable(
+              onDragUpdate: (update) => setState(
+                () => offset = Offset(
+                  offset.dx + update.delta.dx,
+                  offset.dy + update.delta.dy,
+                ),
+              ),
+              onDragEnd: (details) => ref
+                  .read(
+                    pathRecordingMenuUiOffsetProvider.notifier,
+                  )
+                  .update(
+                    Offset(
+                      clampDouble(
+                        offset.dx,
+                        0,
+                        widget.constraints.maxWidth - 380,
+                      ),
+                      clampDouble(
+                        offset.dy,
+                        0,
+                        widget.constraints.maxHeight - 350,
+                      ),
+                    ),
+                  ),
+              childWhenDragging: const SizedBox.shrink(),
+              feedback: const Opacity(
+                opacity: 0.7,
+                child: SizedBox(
+                  height: 700,
+                  child: PathRecordingMenu(),
+                ),
+              ),
+              child: const PathRecordingMenu(),
+            ),
+          ),
+        ),
+      );
 }
