+++ lib/src/features/map/views/map_and_gauge_stack_view.dart
@@ -146,348 +146,17 @@ class MapAndGaugeStackView extends ConsumerWidget {
             ),
           ),
           if (ref.watch(showIMUConfigProvider))
-            Consumer(
-              builder: (context, ref, child) {
-                var offset = ref.read(imuConfiguratorUiOffsetProvider);
-                return StatefulBuilder(
-                  builder: (context, setState) => Positioned(
-                    left: clampDouble(
-                      offset.dx,
-                      0,
-                      constraints.maxWidth - 380,
-                    ),
-                    top: clampDouble(offset.dy, 0, constraints.maxHeight - 350),
-                    child: Padding(
-                      padding: const EdgeInsets.all(8),
-                      child: SizedBox(
-                        height: min(
-                          700,
-                          constraints.maxHeight -
-                              clampDouble(
-                                offset.dy,
-                                0,
-                                constraints.maxHeight - 350,
-                              ),
-                        ),
-                        child: LongPressDraggable(
-                          onDragUpdate: (update) => setState(
-                            () => offset = Offset(
-                              offset.dx + update.delta.dx,
-                              offset.dy + update.delta.dy,
-                            ),
-                          ),
-                          onDragEnd: (details) => ref
-                              .read(
-                                imuConfiguratorUiOffsetProvider.notifier,
-                              )
-                              .update(
-                                Offset(
-                                  clampDouble(
-                                    offset.dx,
-                                    0,
-                                    constraints.maxWidth - 380,
-                                  ),
-                                  clampDouble(
-                                    offset.dy,
-                                    0,
-                                    constraints.maxHeight - 350,
-                                  ),
-                                ),
-                              ),
-                          childWhenDragging: const SizedBox.shrink(),
-                          feedback: const Opacity(
-                            opacity: 0.7,
-                            child: SizedBox(
-                              height: 700,
-                              child: ImuConfigurator(),
-                            ),
-                          ),
-                          child: const ImuConfigurator(),
-                        ),
-                      ),
-                    ),
-                  ),
-                );
-              },
-            ),
+            DraggableImuConfigurator(constraints: constraints),
           if (ref.watch(showSteeringHardwareConfigProvider))
-            Consumer(
-              builder: (context, ref, child) {
-                var offset =
-                    ref.read(steeringHardwareConfiguratorUiOffsetProvider);
-                return StatefulBuilder(
-                  builder: (context, setState) => Positioned(
-                    left: clampDouble(
-                      offset.dx,
-                      0,
-                      constraints.maxWidth - 405,
-                    ),
-                    top: clampDouble(offset.dy, 0, constraints.maxHeight - 350),
-                    child: Padding(
-                      padding: const EdgeInsets.all(8),
-                      child: SizedBox(
-                        height: min(
-                          700,
-                          constraints.maxHeight -
-                              clampDouble(
-                                offset.dy,
-                                0,
-                                constraints.maxHeight - 350,
-                              ),
-                        ),
-                        child: LongPressDraggable(
-                          onDragUpdate: (update) => setState(
-                            () => offset = Offset(
-                              offset.dx + update.delta.dx,
-                              offset.dy + update.delta.dy,
-                            ),
-                          ),
-                          onDragEnd: (details) => ref
-                              .read(
-                                steeringHardwareConfiguratorUiOffsetProvider
-                                    .notifier,
-                              )
-                              .update(
-                                Offset(
-                                  clampDouble(
-                                    offset.dx,
-                                    0,
-                                    constraints.maxWidth - 405,
-                                  ),
-                                  clampDouble(
-                                    offset.dy,
-                                    0,
-                                    constraints.maxHeight - 350,
-                                  ),
-                                ),
-                              ),
-                          childWhenDragging: const SizedBox.shrink(),
-                          feedback: const Opacity(
-                            opacity: 0.7,
-                            child: SizedBox(
-                              height: 700,
-                              child: SteeringHardwareConfigurator(),
-                            ),
-                          ),
-                          child: const SteeringHardwareConfigurator(),
-                        ),
-                      ),
-                    ),
-                  ),
-                );
-              },
-            ),
+            DraggableSteeringHardwareConfigurator(constraints: constraints),
           if (ref.watch(showAutosteeringParameterConfigProvider))
-            Consumer(
-              builder: (context, ref, child) {
-                var offset = ref.read(autosteeringConfiguratorUiOffsetProvider);
-                return StatefulBuilder(
-                  builder: (context, setState) => Positioned(
-                    left: clampDouble(
-                      offset.dx,
-                      0,
-                      constraints.maxWidth - 380,
-                    ),
-                    top: clampDouble(offset.dy, 0, constraints.maxHeight - 390),
-                    child: Padding(
-                      padding: const EdgeInsets.all(8),
-                      child: SizedBox(
-                        height: min(
-                          385,
-                          constraints.maxHeight -
-                              clampDouble(
-                                offset.dy,
-                                0,
-                                constraints.maxHeight - 390,
-                              ),
-                        ),
-                        child: LongPressDraggable(
-                          onDragUpdate: (update) {
-                            setState(
-                              () => offset = Offset(
-                                offset.dx + update.delta.dx,
-                                offset.dy + update.delta.dy,
-                              ),
-                            );
-                          },
-                          onDragEnd: (details) => ref
-                              .read(
-                                autosteeringConfiguratorUiOffsetProvider
-                                    .notifier,
-                              )
-                              .update(
-                                Offset(
-                                  clampDouble(
-                                    offset.dx,
-                                    0,
-                                    constraints.maxWidth - 380,
-                                  ),
-                                  clampDouble(
-                                    offset.dy,
-                                    0,
-                                    constraints.maxHeight - 390,
-                                  ),
-                                ),
-                              ),
-                          childWhenDragging: const SizedBox.shrink(),
-                          feedback: const Opacity(
-                            opacity: 0.7,
-                            child: SizedBox(
-                              height: 385,
-                              child: AutosteeringParameterConfigurator(),
-                            ),
-                          ),
-                          child: const SizedBox(
-                            height: 385,
-                            child: AutosteeringParameterConfigurator(),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                );
-              },
+            DraggableAutosteeringParameterConfigurator(
+              constraints: constraints,
             ),
           if (ref.watch(showPathRecordingMenuProvider))
-            Consumer(
-              builder: (context, ref, child) {
-                var offset = ref.read(pathRecordingMenuUiOffsetProvider);
-                return StatefulBuilder(
-                  builder: (context, setState) => Positioned(
-                    left: clampDouble(
-                      offset.dx,
-                      0,
-                      constraints.maxWidth - 380,
-                    ),
-                    top: clampDouble(offset.dy, 0, constraints.maxHeight - 350),
-                    child: Padding(
-                      padding: const EdgeInsets.all(8),
-                      child: SizedBox(
-                        height: min(
-                          700,
-                          constraints.maxHeight -
-                              clampDouble(
-                                offset.dy,
-                                0,
-                                constraints.maxHeight - 350,
-                              ),
-                        ),
-                        child: LongPressDraggable(
-                          onDragUpdate: (update) => setState(
-                            () => offset = Offset(
-                              offset.dx + update.delta.dx,
-                              offset.dy + update.delta.dy,
-                            ),
-                          ),
-                          onDragEnd: (details) => ref
-                              .read(
-                                pathRecordingMenuUiOffsetProvider.notifier,
-                              )
-                              .update(
-                                Offset(
-                                  clampDouble(
-                                    offset.dx,
-                                    0,
-                                    constraints.maxWidth - 380,
-                                  ),
-                                  clampDouble(
-                                    offset.dy,
-                                    0,
-                                    constraints.maxHeight - 350,
-                                  ),
-                                ),
-                              ),
-                          childWhenDragging: const SizedBox.shrink(),
-                          feedback: const Opacity(
-                            opacity: 0.7,
-                            child: SizedBox(
-                              height: 700,
-                              child: PathRecordingMenu(),
-                            ),
-                          ),
-                          child: const PathRecordingMenu(),
-                        ),
-                      ),
-                    ),
-                  ),
-                );
-              },
-            ),
+            DraggablePathRecordingMenu(constraints: constraints),
           if (ref.watch(showNudgingControlsProvider))
-            Consumer(
-              builder: (context, ref, child) {
-                const height = 325.0;
-                var offset = ref.read(nudgingControlsUiOffsetProvider);
-                return StatefulBuilder(
-                  builder: (context, setState) => Positioned(
-                    left: clampDouble(
-                      offset.dx,
-                      0,
-                      constraints.maxWidth - 380,
-                    ),
-                    top: clampDouble(
-                      offset.dy,
-                      0,
-                      constraints.maxHeight - height,
-                    ),
-                    child: Padding(
-                      padding: const EdgeInsets.all(8),
-                      child: SizedBox(
-                        height: min(
-                          height,
-                          constraints.maxHeight -
-                              clampDouble(
-                                offset.dy,
-                                0,
-                                constraints.maxHeight - height,
-                              ),
-                        ),
-                        child: LongPressDraggable(
-                          onDragUpdate: (update) {
-                            setState(
-                              () => offset = Offset(
-                                offset.dx + update.delta.dx,
-                                offset.dy + update.delta.dy,
-                              ),
-                            );
-                          },
-                          onDragEnd: (details) => ref
-                              .read(
-                                nudgingControlsUiOffsetProvider.notifier,
-                              )
-                              .update(
-                                Offset(
-                                  clampDouble(
-                                    offset.dx,
-                                    0,
-                                    constraints.maxWidth - 380,
-                                  ),
-                                  clampDouble(
-                                    offset.dy,
-                                    0,
-                                    constraints.maxHeight - height,
-                                  ),
-                                ),
-                              ),
-                          childWhenDragging: const SizedBox.shrink(),
-                          feedback: const Opacity(
-                            opacity: 0.7,
-                            child: SizedBox(
-                              height: height,
-                              child: NudgingControls(),
-                            ),
-                          ),
-                          child: const SizedBox(
-                            height: height,
-                            child: NudgingControls(),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                );
-              },
-            ),
+            DraggableNudgingControls(constraints: constraints),
         ],
       ),
     );
